requirements.txt
google-genai>=0.8.0
streamlit>=1.35.0
gspread>=6.0.0
google-auth>=2.29.0